+++ builder_pkgs/mockito/lib/annotations.dart
@@ -71,23 +71,39 @@ class MockSpec<T> {
 
   final bool returnNullOnMissingStub;
 
+  final Set<Symbol> unsupportedMembers;
+
   final Map<Symbol, Function> fallbackGenerators;
 
   /// Constructs a custom mock specification.
   ///
   /// Specify a custom name with the [as] parameter.
   ///
-  /// If [returnNullOnMissingStub] is true, the mock class will return `null`
-  /// when a method is called and no stub could be found. This may result in a
-  /// runtime error, if the return type of the method is non-nullable.
+  /// If [returnNullOnMissingStub] is true, a real call to a mock method will
+  /// return `null` when no stub is found. This may result in a runtime error,
+  /// if the return type of the method is non-nullable.
+  ///
+  /// If the class-to-mock has a member with a non-nullable unknown return type
+  /// (such as a type variable, `T`), then mockito cannot generate a valid
+  /// override member, unless the member is specified in [unsupportedMembers],
+  /// or a fallback implementation is given in [fallbackGenerators].
+  ///
+  /// If [unsupportedMembers] contains the name of each such member, the mock
+  /// class is generated with an override, with throws an exception, for each
+  /// member with a non-nullable unknown return type. Such an override cannot be
+  /// used with the mockito stubbing and verification APIs, but makes the mock
+  /// class a valid implementation of the class-to-mock.
   ///
   /// Each entry in [fallbackGenerators] specifies a mapping from a method name
-  /// to a function, with the same signature as the method. This function will
-  /// be used to generate fallback values when a non-null value needs to be
-  /// returned when stubbing or verifying.
+  /// to a function, with the same signature as the method. This function is
+  /// used to generate fallback values when a non-null value needs to be
+  /// returned when stubbing or verifying. A fallback value is not ever exposed
+  /// in stubbing or verifying; it is an object that mockito's internals can use
+  /// as a legal return value.
   const MockSpec({
     Symbol? as,
     this.returnNullOnMissingStub = false,
+    this.unsupportedMembers = const {},
     this.fallbackGenerators = const {},
   }) : mockName = as;
 }