+++ builder_pkgs/mockito/test/end2end/generated_mocks_test.dart
@@ -6,11 +6,15 @@ import 'foo.dart';
 import 'foo_sub.dart';
 import 'generated_mocks_test.mocks.dart';
 
-T dummyMethod<T>() => [1, 1.5].whereType<T>().first!;
+T returnsTypeVariableShim<T>() => [1, 1.5].whereType<T>().first!;
 
-T dummyBoundedMethod<T extends num?>() => [1, 1.5].whereType<T>().first!;
+T returnsBoundedTypeVariableShim<T extends num?>() =>
+    [1, 1.5].whereType<T>().first!;
 
-T dummyMethodTwo<T, U>() => [1, 1.5].whereType<T>().first!;
+T returnsTypeVariableFromTwoShim<T, U>() => [1, 1.5].whereType<T>().first!;
+
+T typeVariableFieldShim<T>() =>
+    throw UnsupportedError('typeVariableField cannot be used');
 
 @GenerateMocks([
   Foo,
@@ -19,10 +23,17 @@ T dummyMethodTwo<T, U>() => [1, 1.5].whereType<T>().first!;
 ], customMocks: [
   MockSpec<Foo>(as: #MockFooRelaxed, returnNullOnMissingStub: true),
   MockSpec<Bar>(as: #MockBarRelaxed, returnNullOnMissingStub: true),
-  MockSpec<Baz>(as: #MockBaz, fallbackGenerators: {
-    #returnsTypeVariable: dummyMethod,
-    #returnsBoundedTypeVariable: dummyBoundedMethod,
-    #returnsTypeVariableFromTwo: dummyMethodTwo,
+  MockSpec<Baz>(as: #MockBazWithUnsupportedMembers, unsupportedMembers: {
+    #returnsTypeVariable,
+    #returnsBoundedTypeVariable,
+    #returnsTypeVariableFromTwo,
+    #typeVariableField,
+  }),
+  MockSpec<Baz>(as: #MockBazWithFallbackGenerators, fallbackGenerators: {
+    #returnsTypeVariable: returnsTypeVariableShim,
+    #returnsBoundedTypeVariable: returnsBoundedTypeVariableShim,
+    #returnsTypeVariableFromTwo: returnsTypeVariableFromTwoShim,
+    #typeVariableField: typeVariableFieldShim,
   }),
 ])
 void main() {
@@ -157,11 +168,27 @@ void main() {
     });
   });
 
+  group('for a generated mock using unsupportedMembers', () {
+    late Baz baz;
+
+    setUp(() {
+      baz = MockBazWithUnsupportedMembers();
+    });
+
+    test('a real method call throws', () {
+      expect(() => baz.returnsTypeVariable(), throwsUnsupportedError);
+    });
+
+    test('a real getter call (or field access) throws', () {
+      expect(() => baz.typeVariableField, throwsUnsupportedError);
+    });
+  });
+
   group('for a generated mock using fallbackGenerators,', () {
     late Baz baz;
 
     setUp(() {
-      baz = MockBaz();
+      baz = MockBazWithFallbackGenerators();
     });
 
     test('a method with a type variable return type can be called', () {